Bumped to version to 2.0: checking xml * StepsLogger.java: made info() public * BonusSiteMapLinksTest.java: renamed to SitemapTest * PageSource.java: moved to SitemapTest directly and removed * SitemapTest.java: unified for any sitemap, taking xml sitemap source instead of html * pom.xml: resolved slf4j console errors by import slf4j 1.7.21 * structure: changed from com.qiwi.bonus.tests.api to sitemap.tests
